styles for new done. styles for incoming done / notes change colors depending on how long they been created...
