+++ my-app/src/components/MenuSelector/MenuElement.css
@@ -4,8 +4,8 @@
     flex-direction: column;
 
     margin: 10px; 
-    width: 150px; 
-    height: 150px; 
+    width: 170px; 
+    height: 170px; 
 
     align-items: center;
     
@@ -24,14 +24,14 @@
     flex-direction: column;
 
     margin: 10px; 
-    width: 150px; 
-    height: 150px; 
+    width: 170px; 
+    height: 170px; 
 
     align-items: center;
     
     border-radius: 25px; 
 
-    background: #F6DF8D;
+    background: #FFCF87;
 
     font-family: 'Montserrat', sans-serif; 
 
@@ -39,8 +39,8 @@
 }
 
 img{
-    width: 90px; 
-    height: 90px; 
+    width: 60%; 
+    height: 60%;  
 
     margin: 5px 0px 0px 0px; 
 }
@@ -48,12 +48,12 @@ img{
 h2{
     margin: 0px 0px 0px 0px; 
 
-    font-size: 15px; 
+    font-size: 17px; 
     font-weight: bold;
 }
 
 p{
     margin: 5px 0px 10px 0px; 
 
-    font-size: 14px; 
+    font-size: 17px; 
 }
