+++ my-app/src/components/MenuSelector/ToggleMenu.css
@@ -1,24 +1,18 @@
-.container{
+.toggle-container{
 
     display: flex; 
     align-items: center; 
-    width: 350px;
+    width: 400px;
+    padding-bottom: 20px; 
     height: auto; 
     justify-content: space-between;
     
 }
 
 button.toggle{
-    
-    font-family: 'Montserrat', sans-serif;
-    font-size: 22px;
-    width: 170px; 
-    height: 50px; 
+    font-size: 25px;
+    width: 192px;  
+    height: 58px; 
     border: 0; 
     border-radius: 16px;
-
-}
-
-button.toggle:focus{
-    font-weight: bold; 
 }
