+++ my-app/src/components/Navbar/BMenu.css
@@ -107,7 +107,7 @@ into a crossmark. */
   padding-top: 125px;
   text-align: center;
   
-  background: #ffb13b;
+  background: #B8AAE0;
   list-style-type: none;
   -webkit-font-smoothing: antialiased;
   /* to stop flickering of text in safari */
