+++ my-app/src/components/Navbar/Navbar.css
@@ -1,7 +1,7 @@
 .topbar {
     height: 100px;
     width: 100%;  
-    background-color: #ffb13b;
+    background-color: #B8AAE0;
 
     margin-top: -21px;
