+++ my-app/src/components/Navbar/Navbar.js
@@ -4,7 +4,7 @@ import './Navbar.css'
 function Navbar({section}) {
     return(
         <div className= "topbar">
-            <h1>Burguer Queen / {section} </h1>
+            <h1>Burguer Queen {section} </h1>
         </div>
     )
 }
