+++ my-app/src/components/Ready/index.js
@@ -1,0 +1,9 @@
+import React, { useState, useEffect} from 'react'
+
+function Ready () {
+    return(
+        <div> Hola desde Ready! </div>
+    )
+}
+
+export default Ready;
