+++ my-app/src/index.css
@@ -1,11 +1,8 @@
 body {
 	margin: 0;
-	font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-	  'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-	  sans-serif;
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
-	background-color: #FFEBCC;
+	background-color: #F2F2F2; 
   }
   
   code {
@@ -15,6 +12,8 @@ body {
 
   *, html, body {
 	box-sizing: border-box;
+	font-family: 'Montserrat', sans-serif; 
+	color: #170000 !important;
   }
 
   body {
